Fix exponential interpolation
For some reason, the exponential interpolation was doing some weird
subtractions. These subtractions broke the exponential interpolation
such that, with a base of 2, a reduction of zoom level by 1 wouldn't
yield a precise size reduction of factor 2.

This can be reproduced with a circle of sufficient size (e.g. 1000km)
and a function of the form:

    {
        type: 'exponential',
        stops: [[0, 0], [20, 1024]],
        base: 2
    }

I expect the circle to maintain a fixed size relative to the map, i.e.
the circle border should always lie on the same coordinates. Currently,
due to these subtractions, the circle increases in relative size with
reducing zoom level.

This commit removes the unnecessary subtraction, yielding a smooth
division by the base in the example above, such that the circle no
longer changes in size when modifying the zoom level.